requirements.txt
torch>=2.1
torchvision>=0.16
timm>=1.0
gradio>=4.0
numpy
Pillow
